1st version of patcher which is functionnal

+++ src/patcher.py
@@ -19,46 +19,61 @@ import shutil
 
 import resource
 import yaml
-from mmap import mmap
+import mmap
 
 import meta
 
 try:
     PAGESIZE = resource.getpagesize()
+    print("page size is: "+str(PAGESIZE))
 except NameError:
     print("page size issue?")
     exit(1)
 
 
-def overwrite(fileobj, start, end, newbytes):
+def overwrite(fileobj, start: int, end: int, newbytes: bytes):
     startoffset, startremainder = divmod(start, PAGESIZE)
     offset = startoffset * PAGESIZE
     endoffset, endremainder = divmod(end, PAGESIZE)
     length = (endoffset + 1) * PAGESIZE - offset
-    data = mmap.mmap(fileobj.fileno(), offset=offset, length=length,
-                     access=mmap.ACCESS_WRITE)
+    data = mmap(fileobj.fileno(), offset=offset, length=length)
     data[startremainder:startremainder + end - start] = newbytes
 
+def reverse_control_codes(value):
+    value = value.replace("#PN#".encode('shift_jisx0213'), b'\x13\x07')
+    value = value.replace('#0A#'.encode('shift_jisx0213'), b'\x0A')
+    value = value.replace('#07#'.encode('shift_jisx0213'), b'\x07')
+    return value
 
 # copy iso locally, then edit bytes
 
 OUTPUT_DIR = "./"
 PATCHED_DT0 = os.path.join(OUTPUT_DIR, os.path.basename(meta.DT0))
 
-print("copying DTO")
-shutil.copy2(meta.DT0, PATCHED_DT0)
+print("copying from " + meta.DT0 + " to " + PATCHED_DT0)
+if not os.path.exists(meta.DT0):
+    print("expected file not found: "+meta.DT0)
+    exit(1)
+try:
+    shutil.copy2(meta.DT0, PATCHED_DT0)
+except IOError as err:
+    print("DTO file copy failed: "+PATCHED_DT0)
+    exit(1)
 
 print("patching DT0")
-with open(PATCHED_DT0, 'wb') as out:
-    with open("../translations/scenario.yaml", 'rb') as scen:
+with open(PATCHED_DT0, mode="r+") as file_obj:
+    DT0 = mmap.mmap(file_obj.fileno(), length=0, access=mmap.ACCESS_WRITE)
+    with open("./translations/scenario_test.yaml", 'rb') as scen:
         scenario_data = yaml.safe_load(scen)
-        print(scenario_data)
         for block in scenario_data:
             print(str(scenario_data[block]["offsets"]["start"]))
             if scenario_data[block]["text"]["translated"] is not None:
-                overwrite(
-                    out,
-                    scenario_data[block]["offsets"]["start"],
-                    scenario_data[block]["offsets"]["end"],
-                    scenario_data[block]["text"]["translated"]
-                )
+                ori = scenario_data[block]["text"]["translated"]
+                ori2 = ori.encode('shift_jisx0213')
+                ori3 = reverse_control_codes(ori2)
+                start=int(scenario_data[block]["offsets"]["start"], 16)
+                end=int(scenario_data[block]["offsets"]["end"], 16)
+                max_size =int(scenario_data[block]["offsets"]["byte_length"])
+                if end-start <= max_size:
+                    DT0[start: start+len(ori3)] = ori3
+            DT0.flush()
